Move SvcCache and DcCache into their own files.
The consul namer is a little difficult to reason about as one big intermingled
class. We begin to simplify this by moving DcCache and SvcCache into their own
package-local classes.
